% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ATAC_assay.R
\name{add_ATAC_assay}
\alias{add_ATAC_assay}
\title{Add ATAC Assay to Seurat Object}
\usage{
add_ATAC_assay(seurat_obj, atac_assay, assay_name = "ATAC")
}
\arguments{
\item{seurat_obj}{A Seurat object containing RNA data (must have an "RNA" assay).}

\item{atac_assay}{A Seurat object or assay matrix containing ATAC data.}

\item{assay_name}{Character string specifying the name for the ATAC assay in the Seurat object.
Default is "ATAC".}
}
\value{
Returns the updated Seurat object with the ATAC assay added.
}
\description{
This function adds an ATAC assay to an existing RNA Seurat object, aligning cell barcodes
between the RNA and ATAC datasets. It identifies common cells between the two assays,
subsets both objects to these common cells, and adds the ATAC assay to the Seurat object.
}
\details{
The function performs the following steps:
\enumerate{
\item Identifies common cells between RNA and ATAC datasets
\item Subsets both objects to these common cells
\item Verifies cell order alignment after subsetting
\item Adds the ATAC assay to the Seurat object
\item Sets the default assay back to "RNA"
}
}
\note{
\itemize{
\item The function requires that both RNA and ATAC objects have the same cell barcodes
\item Cell order must be consistent between the two datasets
\item The RNA assay is set as the default assay after adding the ATAC assay
}
}
\examples{
# Basic usage
# Assuming 'rna_seurat' is your RNA Seurat object and 'atac_data' is your ATAC assay
updated_seurat <- add_ATAC_assay(rna_seurat, atac_data)

# With custom assay name
updated_seurat <- add_ATAC_assay(rna_seurat, atac_data, assay_name = "ATAC_peaks")

}
\seealso{
\code{\link{DefaultAssay}} for managing default assays in Seurat objects
}
