% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_qc_plots_features.R
\name{seurat_qc_plots_features}
\alias{seurat_qc_plots_features}
\title{Generate all QC plots for Seurat object features                                                                                            \link{112/1808}}
\usage{
seurat_qc_plots_features(
  seurat_obj,
  features = NULL,
  feature_files = NULL,
  output_dir = "./",
  dotplot_file = "QC6.markers_provided.DotPlot",
  heatmap_file = "QC6.markers_provided.DoHeatmap",
  vlnplot_prefix = "gene.",
  featureplot_prefix = "gene.",
  ident_var = NULL,
  dotplot_params = list(),
  heatmap_params = list(),
  vlnplot_params = list(),
  featureplot_params = list()
)
}
\arguments{
\item{seurat_obj}{A Seurat object containing the single-cell RNA-seq data.}

\item{features}{Character vector of features (genes) to plot. Optional if feature_files is provided.}

\item{feature_files}{Character vector of file paths containing features to plot. Each file should contain one feature per line.}

\item{output_dir}{Directory path where the plots will be saved. Default is current directory.}

\item{dotplot_file}{Name of the DotPlot output file (without extension). Default is "QC6.markers_provided.DotPlot".}

\item{heatmap_file}{Name of the DoHeatmap output file (without extension). Default is "QC6.markers_provided.DoHeatmap".}

\item{vlnplot_prefix}{Prefix for VlnPlot output files. Default is "gene.".}

\item{featureplot_prefix}{Prefix for FeaturePlot output files. Default is "gene.".}

\item{ident_var}{Character string specifying the cluster identifier variable in metadata.
If NULL, uses the current identification in the Seurat object.}

\item{dotplot_params}{List of parameters to pass to seurat_dotplot_features.}

\item{heatmap_params}{List of parameters to pass to seurat_heatmap_features.}

\item{vlnplot_params}{List of parameters to pass to seurat_violin_feature_plots for VlnPlot.}

\item{featureplot_params}{List of parameters to pass to seurat_violin_feature_plots for FeaturePlot.}
}
\value{
Invisibly returns a list of paths to all saved plot files.
}
\description{
Convenience function that generates all QC plots (DotPlot, DoHeatmap,
VlnPlot, and FeaturePlot) for specified features in a Seurat object. This function
combines the functionality of the individual plot functions.
}
\examples{
                                                                                                                                            
# Basic usage with feature vector                                                                                                                     
seurat_qc_plots_features(my.seurat, features = c("PECAM1", "CDH5"))                                                                                   
                                                                                                                                                      
# With custom output filenames                                                                                                                        
seurat_qc_plots_features(my.seurat, features = my_genes,                                                                                              
                        dotplot_file = "my_dotplot",                                                                                                  
                        heatmap_file = "my_heatmap",                                                                                                  
                        vlnplot_prefix = "vln_",                                                                                                      
                        featureplot_prefix = "feat_")                                                                                                 

}
